+++ app/src/main/java/com/aspa/aspa/core/constants/enums/Provider.kt
@@ -1,4 +1,4 @@
-package com.aspa.aspa.model
+package com.aspa.aspa.core.constants.enums
 
 enum class Provider() {
     GOOGLE, KAKAO, NAVER
