+++ app/src/main/java/com/aspa/aspa/data/repository/AuthRepository.kt
@@ -1,7 +1,7 @@
 package com.aspa.aspa.data.repository
 
 import com.aspa.aspa.data.dto.UserProfileDto
-import com.aspa.aspa.model.Provider
+import com.aspa.aspa.core.constants.enums.Provider
 
 interface AuthRepository {
